requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==6.0.2
djangorestframework==3.16.1
idna==3.11
itypes==1.2.0
Jinja2==3.1.6
Markdown==3.10.2
MarkupSafe==3.0.3
numpy==2.4.2
pandas==3.0.1
pyarrow==23.0.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
requests==2.32.5
setuptools==80.10.2
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3